+++ src/inmanta/protocol/rest/server.py
@@ -21,6 +21,7 @@ import ssl
 import uuid
 from asyncio import CancelledError
 from collections import defaultdict
+from json import JSONDecodeError
 from typing import Dict, List, MutableMapping, Optional, Union
 
 import tornado
@@ -100,6 +101,8 @@ class RESTHandler(tornado.web.RequestHandler):
             return common.json_encode(body)
         if content_type == common.HTML_CONTENT:
             return body.encode(common.HTML_ENCODING)
+        if content_type == common.HTML_CONTENT_WITH_UTF8_CHARSET:
+            return body.encode(common.UTF8_ENCODING)
         return body
 
     async def _call(self, kwargs: Dict[str, str], http_method: str, call_config: common.UrlMethod) -> None:
@@ -137,6 +140,11 @@ class RESTHandler(tornado.web.RequestHandler):
                     self.respond(result.body, result.headers, result.status_code)
                 else:
                     raise exceptions.UnauthorizedException("Access to this resource is unauthorized.")
+
+            except JSONDecodeError as e:
+                error_message = f"The request body couldn't be decoded as a JSON: {e}"
+                LOGGER.info(error_message, exc_info=True)
+                self.respond({"message": error_message}, {}, 400)
 
             except ValueError:
                 LOGGER.exception("An exception occured")